11/25

+++ database/migrations/2021_11_20_024816_create_tasks_table.php
@@ -14,7 +14,7 @@ class CreateTasksTable extends Migration
     public function up()
     {
         Schema::create('tasks', function (Blueprint $table) {
-            $table->increments('id');
+            $table->bigInteger('id');
             $table->integer('folder_id')->unsigned();
             $table->string('title', 100);
             $table->date('due_date');
